Add makefile and refactor test code
