+++ test/scripts/publish_message.py
